remove export from page - causing it to error out
