+++ components/ParseReview/ParseReviewStars.tsx
@@ -3,8 +3,8 @@ import { FilledStarICon, UnfilledStarIcon } from "@/public/SVGs";
 
 export default function ParseReviewStars({ review }: { review: number }) {
   const stars: ReactNode[] = [];
-  for (let i = 0; i <= 5; i += 1) {
-    if (i <= review) {
+  for (let i = 0; i < 5; i += 1) {
+    if (i < review) {
       stars.push(
         <Fragment key={i}>
           <FilledStarICon />
